refactor(T11202): remove versioned attachment store module
Fold the llmtxt mirror API into the canonical attachment-store module and update callers/tests to use non-versioned names.

Task: T11202
